backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
biopython==1.82
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
openai==1.12.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
